// API_backup/api.ts
export interface PostProfileResponse {
    /**
     * 
     * @type {string}
     * @memberof PostProfileResponse
     */
    'userId'?: string;
    /**
     * 
     * @type {string}
     * @memberof PostProfileResponse
     */
    'imageUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostProfileResponse
     */
    'username'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostProfileResponse
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostProfileResponse
     */
    'lastName'?: string | null;
}
/**
 * 
 * @export
 * @interface PostResponse
 */
export interface PostResponse {
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    'postId'?: string;
    /**
     * 
     * @type {PostProfileResponse}
     * @memberof PostResponse
     */
    'profile'?: PostProfileResponse;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    'content'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PostResponse
     */
    'createdAt'?: string;
}